Make registrationsi simpler with builders
The computedattribute and dtomapping registrations have parameters that
are easily switchable, and that can lead to confusions.
Since java lacks named parameters, it can happen that a user
accidentally swaps the entity and dto in the mapping registration, or
swapping the entity and the attribute type accidentally in the computed
attribute registration.

With the added step builder pattern it is more straightforward what
attribute is required, and the steps allow enforcing that all required
attributes are set.